add docs folder structure and external and util modules
