+++ docs/index.md
@@ -1,0 +1,8 @@
+# Documentation
+
+This is the documentation of the E-Bill backend
+
+## Contents
+
+* [Configuration](configuration.md)
+
